# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mashalliance API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Registration User",
                "description": "Registration User",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "username"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "password"
                                    },
                                    "confirm_password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "confirm_password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth::user"
                ],
                "summary": "Login User",
                "description": "Login User",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "username"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}